+++ common/src/main/java/mb/common/function/Function2.java
@@ -1,0 +1,8 @@
+package mb.common.function;
+
+@FunctionalInterface
+public interface Function2<T1, T2, R> extends
+    Function2Throwing1<T1, T2, R, RuntimeException>,
+    Function2Throwing2<T1, T2, R, RuntimeException, RuntimeException> {
+    @Override R apply(T1 t1, T2 t2);
+}