+++ common/src/main/java/mb/common/result/ThrowingFunction.java
@@ -1,6 +1,8 @@
 package mb.common.result;
 
+import mb.common.function.Function1Throwing1;
+
 @FunctionalInterface
-public interface ThrowingFunction<T, R, E extends Exception> {
-    R apply(T t) throws E;
+public interface ThrowingFunction<T, R, E extends Exception> extends Function1Throwing1<T, R, E> {
+
 }