+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project are documented in this file, based on [Keep 
 
 
 ## [Unreleased]
+### Added
+- `mb.common.function.Function[1-4]`
+- `mb.common.function.Function[1-4]Throwing[1-2]`
+
+### Changed
+- `mb.common.result.ThrowingFunction` extends `Function1Throwing1`.
 
 
 ## [0.10.3] - 2022-04-04